refactor: add placeholders for all API routes
