+++ lib/api/error.ts
