--- lib/api/time-in-availability.ts
@@ -1,8 +1,0 @@
-export default function timeInAvailability(
-  req: Req,
-  res: Res,
-  people: UserWithRoles[],
-  time: Timeslot
-): boolean {
-  return false;
-}
