+++ pages/api/requests/index.ts
@@ -1,8 +1,10 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { ApiError } from 'lib/model';
-import createRequest, { CreateRequestRes } from 'lib/api/create-request';
-import listRequests, { ListRequestsRes } from 'lib/api/list-requests';
+import { APIError } from 'lib/api/error';
+import createRequest, {
+  CreateRequestRes,
+} from 'lib/api/routes/requests/create';
+import listRequests, { ListRequestsRes } from 'lib/api/routes/requests/list';
 
 /**
  * GET - Lists the user's, their children's, or their org's requests.
@@ -10,9 +12,9 @@ import listRequests, { ListRequestsRes } from 'lib/api/list-requests';
  *
  * Requires a JWT owned by a request person, their parents, or their org admin.
  */
-export default async function matches(
+export default async function requests(
   req: NextApiRequest,
-  res: NextApiResponse<ListRequestsRes | CreateRequestRes | ApiError>
+  res: NextApiResponse<ListRequestsRes | CreateRequestRes | APIError>
 ): Promise<void> {
   switch (req.method) {
     case 'GET': // List the user's, their children's, and their org's requests.
